!fixme In-progress work on list publisher
