+++ scripts/avail-ext-doc/Dockerfile
@@ -1,7 +1,7 @@
 FROM python:3.6
 
-RUN pip install Jinja2==2.10 GitPython==2.1.8 PyGithub>=1.36
+COPY . /app
+WORKDIR /app
+RUN pip install -r requirements.txt 
 
-ADD main.py /
-
-CMD python /main.py
+CMD python ./main.py
