+++ scripts/avail-ext-doc/requirements.txt
@@ -1,0 +1,3 @@
+Jinja2==2.10
+GitPython==2.1.8
+PyGithub>=1.36
